requirements.txt
google-generativeai
python-dotenv
streamlit
pillow
langchain-google-genai
google-genai